refactor(foundation): move WebSocket/MQTT/Kafka topic schema types (A21)
Extend `mockforge-foundation::protocol_contract_types` with the
per-protocol topic/message schema data types so consumers can
reference them without pulling in deprecated core modules.

**Moved to foundation:**
- WebSocketMessageType + MessageDirection (from websocket_contract.rs)
- MqttTopicSchema (from mqtt_kafka_contracts.rs)
- KafkaTopicSchema, TopicSchema, SchemaFormat, EvolutionRules (from
  mqtt_kafka_contracts.rs, including Default impl)

**Stays in core:**
- `WebSocketContract`, `MqttContract`, `KafkaContract`, `GrpcContract`
  impls (they hold compiled `jsonschema::Validator` caches and do the
  actual schema validation)
- `ProtocolContractRegistry`, `compare_contracts`

Core's `websocket_contract.rs` and `mqtt_kafka_contracts.rs`
re-export the data types from foundation; unused serde imports
cleaned up. mockforge-http/handlers/protocol_contracts.rs migrated
to import the schema types from foundation directly while keeping
engine/registry imports on core.

Drops another **51 deprecation warnings** from http (218 → 167).
All 574 http + 1818 core + 42 foundation tests pass; workspace
clippy clean.
