+++ crates/mockforge-core/src/contract_drift/mqtt_kafka_contracts.rs
@@ -11,7 +11,6 @@ use crate::contract_drift::protocol_contracts::{
 };
 use crate::protocol_abstraction::Protocol;
 use jsonschema::{self, Draft, Validator as JSONSchema};
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
@@ -19,22 +18,8 @@ use std::collections::HashMap;
 // MQTT Contract
 // ============================================================================
 
-/// MQTT topic schema definition
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MqttTopicSchema {
-    /// Topic name (supports wildcards + and #)
-    pub topic: String,
-    /// Quality of Service level (0, 1, or 2)
-    pub qos: Option<u8>,
-    /// JSON schema for messages on this topic
-    pub schema: Value,
-    /// Whether messages are retained
-    pub retained: Option<bool>,
-    /// Description of this topic
-    pub description: Option<String>,
-    /// Example message payload
-    pub example: Option<Value>,
-}
+// MqttTopicSchema re-exported from foundation.
+pub use mockforge_foundation::protocol_contract_types::MqttTopicSchema;
 
 /// MQTT contract implementation
 ///
@@ -827,70 +812,10 @@ pub fn diff_mqtt_contracts(
 // Kafka Contract
 // ============================================================================
 
-/// Schema format for Kafka messages
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum SchemaFormat {
-    /// JSON schema format
-    Json,
-    /// Avro schema format
-    Avro,
-    /// Protobuf schema format
-    Protobuf,
-}
-
-/// Kafka topic schema definition
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct KafkaTopicSchema {
-    /// Topic name
-    pub topic: String,
-    /// Key schema (optional - for keyed messages)
-    pub key_schema: Option<TopicSchema>,
-    /// Value schema (required - message payload)
-    pub value_schema: TopicSchema,
-    /// Number of partitions
-    pub partitions: Option<u32>,
-    /// Replication factor
-    pub replication_factor: Option<u16>,
-    /// Description of this topic
-    pub description: Option<String>,
-    /// Evolution rules for schema changes
-    pub evolution_rules: Option<EvolutionRules>,
-}
-
-/// Schema definition for a topic (key or value)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TopicSchema {
-    /// Schema format (JSON, Avro, Protobuf)
-    pub format: SchemaFormat,
-    /// Schema definition (JSON schema, Avro schema JSON, or proto descriptor)
-    pub schema: Value,
-    /// Schema registry ID (if using schema registry)
-    pub schema_id: Option<String>,
-    /// Schema version
-    pub version: Option<String>,
-}
-
-/// Evolution rules for schema changes
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EvolutionRules {
-    /// Allow backward compatible changes (add optional fields, etc.)
-    pub allow_backward_compatible: bool,
-    /// Allow forward compatible changes (remove optional fields, etc.)
-    pub allow_forward_compatible: bool,
-    /// Require explicit version bump for breaking changes
-    pub require_version_bump: bool,
-}
-
-impl Default for EvolutionRules {
-    fn default() -> Self {
-        Self {
-            allow_backward_compatible: true,
-            allow_forward_compatible: false,
-            require_version_bump: true,
-        }
-    }
-}
+// SchemaFormat, KafkaTopicSchema, TopicSchema, EvolutionRules re-exported from foundation.
+pub use mockforge_foundation::protocol_contract_types::{
+    EvolutionRules, KafkaTopicSchema, SchemaFormat, TopicSchema,
+};
 
 /// Kafka contract implementation
 ///
